Contain a complete Vim Configuration V0.1 and an cscope autoload database plugin.
This is is a Simple Vim configuration consistis of the features uesd
in your daily development and code review, just need to copy vimrc
file and the autoload_cscope.vim.
The configuration file has the following features.
-autoindent
-show linenumber
-change colorscheme
-autocomplete the parenthesis
-switch from left to right in a new line
-visual complete command
-immediate show the searche result
-uniform the encoding set
-the directory tree
-keep the cursor always in the middle position
-(Todo)
And one more thing is that used the cscope plugin for the convience
of viewing code. But one tricky thing is that the database can not
be load automaticly. So I use the Vim official plugin to make it
perfect.